# Change how differences in AV rules are generated in sediff