@jwcart2 2019-10-02
I dropped the changes to AVX and TE rules and just changed how AV rules are processed. I now convert the added, removed, and modified rules back into an AVRule class and leave the sediff code itself alone. It also now passes the tests (at least on my machine).