@jwcart2 2019-10-02
I see that I need 2 blank lines after functions. A simple fix. I'll send a new pull request shortly.